requirements.txt
# CVSense Requirements

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Document Processing
python-docx>=0.8.11

# NLP & Text Processing
nltk>=3.8.0

# Feature Extraction & ML
scikit-learn>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Google Drive Integration
gdown>=4.7.0
requests>=2.28.0

# Utilities
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.29.0
plotly>=5.18.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
